# swagger.yaml
openapi: 3.0.0
info:
  title: 用戶認證 API
  description: 提供用戶登入、驗證和記錄相關的 API
  version: 1.0.0

servers:
  - url: http://10.22.94.69:3000
    description: 開發環境

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    StandardResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          description: 響應狀態
        message:
          type: string
          description: 響應訊息
        time:
          type: integer
          description: 時間戳
          
    LoginRequest:
      type: object
      required:
        - EMPID
        - PWD
      properties:
        EMPID:
          type: string
          description: 用戶帳號
        PWD:
          type: string
          description: 用戶密碼
    
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            token:
              type: string
              description: JWT token
            data:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                DeptName:
                  type: string
                Status:
                  type: boolean

paths:
  /user/login:
    post:
      summary: 用戶登入
      description: 驗證用戶憑證並返回 JWT token
      tags:
        - 認證
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 認證失敗
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - example:
                      status: error
                      message: 用戶不存在或帳號密碼錯誤，請重新輸入
                      time: 1234567890
        '500':
          description: 服務器錯誤
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - example:
                      status: error
                      message: 服務暫時無法使用，請稍後再試
                      time: 1234567890

  /user/verify:
    get:
      summary: 驗證用戶 Token
      description: 驗證用戶的 JWT token 是否有效
      tags:
        - 認證
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token 有效
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          DeptName:
                            type: string
        '401':
          description: 未登入
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - example:
                      status: error
                      message: 未登入
                      time: 1234567890
        '403':
          description: 驗證錯誤
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - example:
                      status: error
                      message: 驗證錯誤
                      time: 1234567890

  /user/record:
    post:
      summary: 創建訪問記錄
      description: 記錄用戶的訪問資訊
      tags:
        - 記錄
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ID
                - Name
                - DeptName
                - Time
                - Path
              properties:
                ID:
                  type: string
                Name:
                  type: string
                DeptName:
                  type: string
                Time:
                  type: string
                Path:
                  type: string
      responses:
        '200':
          description: 記錄成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
        '500':
          description: 記錄失敗
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - example:
                      status: error
                      message: 記錄創建失敗
                      time: 1234567890

  /user/record/{st}:
    get:
      summary: 獲取訪問記錄數量
      description: 獲取特定時間後的登入記錄數量
      tags:
        - 記錄
      security:
        - BearerAuth: []
      parameters:
        - name: st
          in: path
          required: true
          description: 起始時間戳記
          schema:
            type: string
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            Count:
                              type: integer
        '500':
          description: 查詢失敗
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - example:
                      status: error
                      message: 記錄查詢失敗
                      time: 1234567890